backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
google-genai==0.1.8
PyPDF2==3.0.1